+++ code/modules/robotics/bot/skullbot.dm
@@ -3,7 +3,7 @@
 /obj/machinery/bot/skullbot
 	name = "skullbot"
 	desc = "A skull on a leg. Useful, somehow. I guess."
-	icon = 'icons/obj/aibots.dmi'
+	icon = 'icons/obj/bots/aibots.dmi'
 	icon_state = "skullbot"
 	layer = 5.0 //TODO LAYER
 	density = 0
@@ -74,11 +74,11 @@
 		return src.explode()
 
 	explode()
+		if(src.exploding) return
+		src.exploding = 1
 		src.on = 0
 		src.visible_message("<span class='combat'><B>[src] blows apart!</B></span>")
-		var/datum/effects/system/spark_spread/s = unpool(/datum/effects/system/spark_spread)
-		s.set_up(3, 1, src)
-		s.start()
+		elecflash(src, radius=1, power=3, exclude_center = 0)
 		qdel(src)
 		return
